% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listing.R
\name{lsos}
\alias{lsos}
\title{List objects in an environment}
\usage{
lsos(..., n = 10)
}
\arguments{
\item{...}{arguments passed to `ls.objects`}

\item{n}{number of rows to return}
}
\value{
a data.frame
}
\description{
List objects in an environment
}
