% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listing.R
\name{make_document}
\alias{make_document}
\title{Create a new Quarto document file.}
\usage{
make_document(filename)
}
\arguments{
\item{filename}{the filename for the new document}
}
\description{
Create a new Quarto document file.
}
