% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfpmisc-package.R
\docType{package}
\name{mfpmisc-package}
\alias{mfpmisc}
\alias{mfpmisc-package}
\title{mfpmisc: Marc's Random Utilities}
\description{
A collection of functions that are useful in your regular R workflow. No function in this package is expected to appear in R scripts. It is intended that this package can be attached (*e.g.* in your .Rprofile) without much risk of introducing irreproducibility into the scripts you write. This is because functions in this package are intended for interactive use, and not for use in scripts.
}
\author{
\strong{Maintainer}: Marc Paterno \email{marc.paternot@gmail.com} (\href{https://orcid.org/0000-0003-0808-8388}{ORCID})

}
\keyword{internal}
